chore: rename benchsdk to @benchsdk/client
Rename the npm package from 'benchsdk' to '@benchsdk/client' to avoid
naming conflicts with the existing 'bench-sdk' package on npm. The
scoped name under the @benchsdk org bypasses npm's similarity check.

Changes:
- packages/benchsdk/package.json: name → @benchsdk/client
- Root package.json: dependency key updated
- All import statements (5 scale files): 'benchsdk' → '@benchsdk/client'
- CI workflows (ci.yml, release.yml): --filter @benchsdk/client
- Dockerfile: --filter @benchsdk/client
- Contract test: package name assertion updated
- Changeset: @benchsdk/client patch
- README: title, install, and import examples updated
- pnpm-lock.yaml regenerated

Directory name packages/benchsdk/ is unchanged (only npm package name).
