+++ benchmarks/scale/sdk-coordinator.ts
@@ -2,8 +2,8 @@ import { config as loadDotenv } from 'dotenv';
 import * as fs from 'node:fs';
 import * as os from 'node:os';
 import { monitorEventLoopDelay } from 'node:perf_hooks';
-import { createBenchmarkClient, defineStep, defineTask, runBenchmarkWorker } from 'benchsdk';
-import type { BenchmarkAssignment, JsonObject, TaskResultRecord, TaskStepRecord } from 'benchsdk';
+import { createBenchmarkClient, defineStep, defineTask, runBenchmarkWorker } from '@benchsdk/client';
+import type { BenchmarkAssignment, JsonObject, TaskResultRecord, TaskStepRecord } from '@benchsdk/client';
 import { getProvider } from './providers.js';
 import { log } from './logger.js';
 import type { FailureClass, MetricsSample, SandboxResult, SandboxResultStatus } from './types.js';
