+++ package.json
@@ -92,6 +92,7 @@
     "@aws-sdk/s3-presigned-post": "^3.1076.0",
     "@aws-sdk/s3-request-presigner": "^3.1076.0",
     "@azure/storage-blob": "^12.33.0",
+    "@benchsdk/client": "workspace:*",
     "@computesdk/anchorbrowser": "^0.2.4",
     "@computesdk/archil": "^0.4.7",
     "@computesdk/beam": "^0.3.0",
@@ -136,7 +137,6 @@
     "@storagesdk/core": "^0.4.2",
     "@tigrisdata/storage": "^3.16.0",
     "@vercel/blob": "^2.5.0",
-    "benchsdk": "workspace:*",
     "computesdk": "^4.1.4",
     "dotenv": "^17.4.2",
     "e2b": "^2.35.0",
