+++ packages/benchsdk/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "benchsdk",
+  "name": "@benchsdk/client",
   "version": "0.2.0",
   "private": false,
   "type": "module",
